CLI compatible with no active themes

+++ cli/BlockClean.php
@@ -4,6 +4,7 @@ use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 class BlockClean extends AcfgbCommand
 {
@@ -22,25 +23,35 @@ class BlockClean extends AcfgbCommand
     protected function command_init()
     {
         if (function_exists('get_template_directory')){
-            $this->print("------ Init ACFGB Clean ------");
 
-            // Delete block dir in theme
-            $this->delete_block_dir_in_theme();
+            $helper = $this->getHelper('question');
+            $question = new ConfirmationQuestion('<comment>!! Are you sure you want to delete all files?? (y/n)</comment>', false);
+            $confirm = $helper->ask($this->input, $this->output, $question);
 
-            // Delete blocks scss file in theme
-            $this->delete_block_scss_file();
+            if ($confirm == 'y' || $confirm == "yes"){
+                $this->print("------ Init ACFGB Clean ------");
 
-            // Delete block cli file in theme
-            $this->delete_block_cli_file();
+                // Delete block dir in theme
+                $this->delete_block_dir_in_theme();
 
-            //$this->delete_blocks_scss_in_main();
-            $this->print(
-                " - IMPORTANT! Remember delete block scss file reference in main.scss",
-                'comment');
-            $this->print(" - IMPORTANT! If you are using custom JS, remember delete JS routes in main.js",
-                'comment');
+                // Delete blocks scss file in theme
+                $this->delete_block_scss_file();
 
-            $this->print($this->default_messages['tasks_ready']);
+                // Delete block cli file in theme
+                $this->delete_block_cli_file();
+
+                //$this->delete_blocks_scss_in_main();
+                $this->print(
+                    " - IMPORTANT! Remember delete block scss file reference in main.scss",
+                    'comment');
+                $this->print(" - IMPORTANT! If you are using custom JS, remember delete JS routes in main.js",
+                    'comment');
+
+                $this->print($this->default_messages['tasks_ready']);
+            }else{
+                $this->print("<comment>Action canceled</comment>. <info>Your files are safe =)</info>", 'comment');
+            }
+
         }else{
             $error = 'WordPress has not been loaded. This command need use get_template_directory().';
             $this->print($error, 'error');
@@ -49,6 +60,7 @@ class BlockClean extends AcfgbCommand
     }
 
     public function delete_block_dir_in_theme(){
+
         $error = $this->fileManager()->delete_dir(
             $this->theme_plugin_dir,
             "Delete acf-gutenberg folder in theme"
